+++ NQueen.UnitTests/Base/TestBasePartOne.cs
@@ -2,17 +2,18 @@
 
 public partial class TestBase(ISolverBackEnd sut)
 {
-    public List<byte[]> ExpectedSolutions { get; set; } = [];
+    public List<int[]> ExpectedSolutions { get; set; } = [];
 
-    public List<byte[]> ActualSolutions { get; set; } = [];
+    public List<int[]> ActualSolutions { get; set; } = [];
 
-    public static List<byte[]> GetExpectedSolutions(byte boardSize, SolutionMode solutionMode) => solutionMode == SolutionMode.Single
-               ? [.. GetExpectedSingleSolution(boardSize)]
-               : solutionMode == SolutionMode.Unique
-               ? [.. GetExpectedUniqueSolutions(boardSize)]
-               : [.. GetExpectedAllSolutions(boardSize)];
+    public static List<int[]> GetExpectedSolutions(int boardSize, SolutionMode solutionMode) =>
+        solutionMode == SolutionMode.Single
+            ? [.. GetExpectedSingleSolution(boardSize)]
+            : solutionMode == SolutionMode.Unique
+            ? [.. GetExpectedUniqueSolutions(boardSize)]
+            : [.. GetExpectedAllSolutions(boardSize)];
 
-    public List<byte[]> GetActualSolutions(byte boardSize, SolutionMode solutionMode) => [.. Sut
+    public List<int[]> GetActualSolutions(int boardSize, SolutionMode solutionMode) => [.. Sut
                .GetResultsAsync(boardSize, solutionMode)
                .Result
                .Solutions
